Add compare histogram

+++ Lab01/lib.h
@@ -15,6 +15,6 @@ using namespace cv;
 #define min_f(a,b,c) (fminf(a,fminf(b,c)))
 #define max_f(a,b,c) (fmaxf(a,fmaxf(b,c)))
 // Chuyển 1 số nằm trong 1 range này vào 1 range khác
-int normalize_num(int oldValue, int oldMin, int oldMax, int newMin, int newMax); 
+double normalize_num(double oldValue, double oldMin, double oldMax, double newMin, double newMax);
 // Chuyển histogram vào 1 range nhất định
 void normalize_hist(const Mat& inputHist, Mat& outputHist, int min, int max);
